requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
babel==2.3.4
beautifulsoup4==4.5.1     # via bs4
bs4==0.0.1
click==6.6                # via flask
flask-cors==3.0.2
flask==0.11.1
itsdangerous==0.24        # via flask
Jinja2==2.8               # via flask
MarkupSafe==0.23          # via jinja2
python-decouple==3.0
pytz==2016.7              # via babel
redis==2.10.5
requests==2.11.1
Six==1.10.0               # via flask-cors
Werkzeug==0.11.11         # via flask
